Tukar int kepada size_t

+++ string.c
@@ -1,5 +1,5 @@
 #include <stdio.h>
-int my_strlen(char *s)
+size_t my_strlen(char *s)
 {
     char *p = s;
